Fixed ability to send TE bucks. Added some front end to request TE bucks

+++ capstone/tenmo-client/src/main/java/com/techelevator/tenmo/model/Transfer.java
@@ -28,7 +28,7 @@ public class Transfer {
         return toUserId;
     }
 
-    public void setToUserId(long toUserId) {
+    public void setToUserId(int toUserId) {
         this.toUserId = toUserId;
     }
